WIP: Reload and start profiling
DISCLAIMER: This is a first experimental commit about an approach that could work
on reload. It is a PoC around a specific approach.

The main idea is that we expose a call on hooks that the app realm could
use to be notified that profiler is ready and recording. Then on profiler's
backend we wait a call on the bridge that the profiler is initialized
and then we let the hook and the app know that profiler is ready.

With that primitive on the hooks side the app could wait before the root
component on the container dom node.

In order for react-devtools to operate they "require" a renderer on the
page so we render a dummy message and wait for profilers signal to mount
the actual app. This technique was promoted by Brian on
https://twitter.com/brian_d_vaughn/status/1058392617405169664. When
profiler is ready the app must unmount the record before mount the new
one in order for store to track the new root and be able to map
snapshots to a "selectedRoot".

PROS:
1. We can handle the async nature of rendering the app and initializing
the profiler
2. The implementation will be actually glue code on existing features. No
need for new mechanisms.
3. For bundles with profiling this could be the default behavior on
render with flags on the profiler side so if you don't enable it the
behavior remains the same as before.

CONS:
1. Needs coordination from React/ReactDOM
2. In case this was the default behavior may add some delay on the first
rendering when using a react bundle with profiling.